Create character store
